Move TelemetryClient property to DialogContainer
